Cleanup after seq schedulers

+++ transport-planners/src/main/java/com/oskopek/transport/planners/sequential/BacktrackRestartWithAroundPathPickupPlanner.java
@@ -73,7 +73,7 @@ public class BacktrackRestartWithAroundPathPickupPlanner extends SequentialRando
     private ImmutablePlanState newStateRecursively(Domain domain, final ImmutablePlanState current,
             final float exploration) {
         if (current.isGoalState()) {
-            double totalTime = current.getTotalTime();
+            int totalTime = current.getTotalTime();
             if (getBestPlanScore() > totalTime) {
                 logger.debug("Found new best plan {} -> {}", getBestPlanScore(), totalTime);
                 savePlanIfBetter(totalTime, new SequentialPlan(current.getAllActionsInList()));
