+++ transport-planners/src/main/java/com/oskopek/transport/planners/sequential/RandomizedRestartPlanner.java
@@ -94,7 +94,7 @@ public class RandomizedRestartPlanner extends SequentialRandomizedPlanner {
             }
 
             // TODO: collapse plan?
-            double totalTime = current.getTotalTime();
+            int totalTime = current.getTotalTime();
             if (getBestPlanScore() > totalTime) {
                 logger.debug("Found new best plan {} -> {}", getBestPlanScore(), totalTime);
                 savePlanIfBetter(totalTime, new SequentialPlan(current.getAllActionsInList()));
