+++ transport-planners/src/main/java/com/oskopek/transport/planners/sequential/RandomizedRestartWithOnPathPickup2Planner.java
@@ -86,7 +86,7 @@ public class RandomizedRestartWithOnPathPickup2Planner extends RandomizedRestart
             }
 
             // TODO: collapse plan?
-            double totalTime = current.getTotalTime();
+            int totalTime = current.getTotalTime();
             if (getBestPlanScore() > totalTime) {
                 logger.debug("Found new best plan {} -> {}", getBestPlanScore(), totalTime);
                 savePlanIfBetter(totalTime, new SequentialPlan(current.getAllActionsInList()));
