+++ transport-planners/src/main/java/com/oskopek/transport/planners/sequential/RandomizedRestartWithOnPathPickupPlanner.java
@@ -77,7 +77,7 @@ public class RandomizedRestartWithOnPathPickupPlanner extends SequentialRandomiz
             }
 
             // TODO: collapse plan?
-            double totalTime = current.getTotalTime();
+            int totalTime = current.getTotalTime();
             if (getBestPlanScore() > totalTime) {
                 logger.debug("Found new best plan {} -> {}", getBestPlanScore(), totalTime);
                 savePlanIfBetter(totalTime, new SequentialPlan(current.getAllActionsInList()));
